Need to sort by pt for most validation steps to succeed
Seems to be required in pp and pythia. Probably would be required in
PbPb if the accepted jet multiplicity was higher. However, embedded
shouldn't be sorted since it's apparently not in AliPhysics. For that
case, we'll have to sort go change it by hand, but since that's rare,
it's better to do the right thing most of the time