Using uproot for scale factors seems to work now!
This started because of a memory issue on macOS with ROOT 6.24.06 which
I wasn't able to resolve. But as far as I can tell, the TProfile issues
are solved upstream, and it gives the same values for the scale factors
as ROOT, so we should be okay to use uproot. I left the testing code in
so that I can confirm this again later if I'm nervous