Require both det and part level to have at least one particle
Otherwise, we can't perform jet finding on that entire array. Better to
throw out the events and not lose the rest of the array! (We only lose
the chance to calculate some sort of efficiency of there being particle
level particles that are fully lost at detector level, which should be
extremely rare. Also, I'm unsure how I would interpret them)