Basic analysis code for time reclustering
This is just a copy with of the substructure, with a modification to the
clustering algorithm for reclustering. That's all we need for now. This
is a lot of copied code, but I'd rather keep the analyses separate
instead of worrying about whether I was breaking one when making changes
for the other