Add option to avoid returning hists (default)
This appears to help keep memory usage down. Otherwise, we have a
copies of the hists per task, and we can't get them out of memory
easily. It can be useful to return them to merge them on the fly, but it
requires some care which we can't really do in a jupyter notebook