Pass the collision system to the analysis function
Although it's not always useful, it frequently is, so I think it's worth
making part of the interface. The alternative is to make the user
responsible, but this is also a bit tricky because it can't really be
specified in the config, so it has to be added dynamically. This is just
simpler, easier, and less error prone