Ensure we don't pass along option types
I think this was hidden before because I never saw the intermediate
arrays (only the final arrays which were converted to np). This
conversion apparently will convert an option to a concrete array
(or presumably fail if there actually are nones is the array). All of
that is to say, there's no real issue once it's converted to np because
of the particular application, but better now to do the explicit right
thing