Explicitly control exploding outputs, allow more skim output types
Although it's really nice to not have to think about exploding skim
fields into separate directories, it's also tricky to get this right. As
an example, consider the mishandling the flat tree skim (ie. exploding
when it wasn't supposed to do so). I could have manually fixed it,
but then I was wrapping the output for no reason other than tricky
into not exploding. Better to just manually control it and be sure I'm
getting it right