Fix out of bounds access for pt dependent tracking eff
This is entirely an edge case, but I did find a detector level track
with pt = 999.49, which is out of bounds. Since the upper bound of the
last bin of the pt dependent tracking efficiency is arbitrary - it's
just supposed to so big that we never hit it. But we did manage to hit
it! :joy: So we push it even higher to 9999.0.