Increase heartbeat timeout on hiccup to save managers, add one retry
Managers seem to be marked lost, but then they send results later (which
are unexpected), which suggests that they are not actually lost, but the
heartbeat got held up for some reason. Seemed to usually resolve within
5 mins, so we give it 10 mins for a buffer.

This is the first time I've seen this behavior, which was seen on hiccup
for central PbPb embedding (i.e. on of the most intensive computations).
Will keep an eye on how the performance evolves (since it may have other
unexpected consequences. e.g. if a manager actually died, parsl may
still send it tasks which it can't respond to, which I imagine can be
disruptive).

Also to this end, enabled one retry of a failed task. If it fails due to
a transient issue such as a manager failure due like in this case, the
task may be successful on a second run. However, if it's due to a bug,
it won't help, and this is slightly wasteful. I think this could be
refined further with a `retry_handler`, but good enough for now.