Add more generator config details in yaml + flexibility for HF tree
This is needed so that we can determine the columns to use and load more
dynamically. Otherwise, it's a huge pain to e.g. use JEWEL.

In the process, I added a new convention for defining generator options,
skim parameters, and generator analysis parameters. The full
implementation of the generation analysis parameters is still WIP