Handle when no reclustered jets are available for MC
If not otherwise handled, it will raise an exception that will bubble up
to the steering, appearing as a failure. This isn't wrong (we can't
recluster if there are no jets...), but it's a bit surprising to see failed
tasks. Better to handle it explicitly and pass on the message, as is
done for other cases (ie. for the one level and three level)