Update fastjet, fix for macOS, support using commit hash
There's a bug in the DO Run II plugins that seems to have appeared due
to more careful checks in recent versions of clang. There's a commit in
the fastjet main repo, but no release, and there's a lot of friction
using those (need additional dependencies on macOS, for example). So
easier to just to take a tagged version.

I updated to newer versions since there was a linking issue with folder
version of fjcontrib (most likely due to newer versions of the macOS
command line tools upgrading clang and the rest).

Note that the commit hash is experimental (I moved back to the tag before
fully testing), but should either work or be very close (as long as
dependencies are appropriately setup. Namely, you need a recent version
of e.g. libtool on macOS)