Use the recoil axis, implement the exclusive selection more properly
By selecting independently before, I think (but am not sure) that it was
giving me incorrect selections where e.g. one jet passed a pt cut and
another passed an eta cut. This new approach is much simpler to work with.

NOTE: My implementation of the "dijet-axis" was nonsense anyway (as I
suspected), so this is better in any case